shared folder refactoring
