+++ src/apps/contentScripts/lobby/extensions/challengeList/ChallengeListModifier.ts
@@ -1,12 +1,12 @@
   // @ts-check
-import { cleanSelector } from '@/shared/StringUtils'
+import { cleanSelector } from '@/shared/plugins/string.setup'
 import $ from 'jquery'
 import { domEntityType } from '../../types/domEntityType'
 import { createApp } from 'vue'
 import KDRComponent from '@/components/KDR.vue'
-import { gcSelectors } from '@/shared/gc/selectors'
+import { gcSelectors } from '@/shared/extras/gc/selectors'
 import serializer from '../../serializer'
-import BrowserStorage from '@/shared/storage'
+import BrowserStorage from '@/shared/tools/storage'
 
 export default class ChallengeListModifier {
