+++ src/apps/contentScripts/lobby/extensions/lobbies/LobbiesModifier.ts
@@ -1,14 +1,14 @@
 // @ts-check
-import { cleanSelector } from '@/shared/StringUtils'
+import { cleanSelector } from '@/shared/plugins/string.setup'
 import $ from 'jquery'
 import { domEntityType } from '../../types/domEntityType'
 import { createApp } from 'vue'
 import KDRComponent from '@/components/KDR.vue'
 import GCCLobbyTitle from '@/components/GCCLobbyTitle.vue'
-import { gcSelectors } from '@/shared/gc/selectors'
+import { gcSelectors } from '@/shared/extras/gc/selectors'
 import serializer from '../../serializer'
 import lobbyFilter from '../filters/lobbyFilter'
-import BrowserStorage from '@/shared/storage'
+import BrowserStorage from '@/shared/tools/storage'
 import Lobby from '../../types/Lobby'
 import GCCLobbyPinner from '@/components/GCCLobbyPinner.vue'
 import LobbyPlayer from '../../types/LobbyPlayer'
