+++ src/apps/contentScripts/team/TeamModifier.ts
@@ -1,13 +1,13 @@
 // @ts-check
-import { cleanSelector } from '@/shared/StringUtils'
+import { cleanSelector } from '@/shared/plugins/string.setup'
 import $ from 'jquery'
 import { domEntityType } from '../lobby/types/domEntityType'
 import { createApp } from 'vue'
 import KDRComponent from '@/components/KDR.vue'
-import { gcSelectors } from '@/shared/gc/selectors'
+import { gcSelectors } from '@/shared/extras/gc/selectors'
 import serializer from '../lobby/serializer'
 import LobbyPlayer from '../lobby/types/LobbyPlayer'
-import BrowserStorage from '@/shared/storage'
+import BrowserStorage from '@/shared/tools/storage'
 
 export default class TeamModifier {
   challenger: any
