+++ src/apps/popup/index.ts
@@ -1,7 +1,7 @@
-import '../../plugins/window.setup'
+import '../../shared/plugins/window.setup'
 import { createApp } from 'vue'
 import PopupApp from './PopupApp.vue'
-import BrowserStorage from '../../shared/storage'
+import BrowserStorage from '../../shared/tools/storage'
 
 (async function() {
   await BrowserStorage.setup()
