+++ src/shared/extras/gc/api/resources/constants/magicNumbers.ts
