+++ src/shared/extras/gc/types/BooleanNumber.ts
