+++ src/shared/extras/gc/types/GCInitialPlayerStats.ts
