+++ src/shared/extras/gc/types/GCPlayerStatsHistory.ts
