+++ src/shared/settings/observer.jquery.js
