+++ src/shared/settings/string.setup.ts
