+++ src/shared/settings/window.setup.ts
