+++ src/shared/tools/analytics/events.ts
