+++ src/shared/tools/analytics/types.ts
