+++ src/shared/tools/health/checkGCSelectors.js
