+++ src/shared/tools/streaker/GCCMarkEmojiMap.ts
